requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
ca-certificates=2022.4.26=haa95532_0
catalogue=2.0.6=py310h5588dad_2
certifi=2022.6.15=py310haa95532_0
cffi=1.15.0=py310h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=7.1.2=pyh9f0ad1d_0
colorama=0.4.5=pyhd8ed1ab_0
cpuonly=2.0=0
cryptography=37.0.1=py310h21b164f_0
cymem=2.0.6=py310h8a704f9_3
cython-blis=0.7.7=py310h2873277_1
dataclasses=0.8=pyhc8e2a94_3
de-dep-news-trf=3.3.0=pypi_0
filelock=3.7.1=pypi_0
freetype=2.10.4=hd328e21_0
huggingface-hub=0.8.1=pypi_0
icc_rt=2019.0.0=h0cc432a_1
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h2bbff1b_0
langcodes=3.3.0=pyhd8ed1ab_0
libffi=3.4.2=hd77b12b_4
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=he0120a3_1
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
markupsafe=2.1.1=py310he2412df_1
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
murmurhash=1.0.7=py310h8a704f9_0
networkx=2.7.1=pyhd3eb1b0_0
numpy=1.22.3=py310h6d2d95c_0
numpy-base=1.22.3=py310h206c741_0
openssl=1.1.1o=h2bbff1b_0
packaging=21.3=pyhd8ed1ab_0
pathy=0.6.1=pyhd8ed1ab_0
pillow=9.0.1=py310hdc2b20a_0
pip=21.2.4=py310haa95532_0
preshed=3.0.6=py310h8a704f9_2
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=py310he2412df_2
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py310haa95532_0
python=3.10.4=hbb2ffb3_0
python_abi=3.10=2_cp310
pytorch=1.11.0=py3.10_cpu_0
pytorch-mutex=1.0=cpu
pyyaml=6.0=pypi_0
regex=2022.6.2=pypi_0
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=1.0.2=py310h4ed8f06_0
scipy=1.7.3=py310h6d2d95c_0
setuptools=61.2.0=py310haa95532_0
shellingham=1.4.0=pyh44b312d_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=pyhd8ed1ab_0
spacy=3.3.0=py310hef0f399_0
spacy-alignments=0.8.5=pypi_0
spacy-legacy=3.0.9=pyhd8ed1ab_0
spacy-loggers=1.0.2=pyhd8ed1ab_0
spacy-transformers=1.1.6=pypi_0
sqlite=3.38.3=h2bbff1b_0
srsly=2.4.3=py310h8a704f9_1
thinc=8.0.17=py310hc781a3c_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.12.1=pypi_0
torchaudio=0.11.0=py310_cpu
torchvision=0.12.0=py310_cpu
tqdm=4.64.0=pyhd8ed1ab_0
transformers=4.19.4=pypi_0
typer=0.4.1=pyhd8ed1ab_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py310haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wasabi=0.9.1=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py310haa95532_0
wincertstore=0.2=py310haa95532_2
xz=5.2.5=h8cc25b3_1
zlib=1.2.12=h8cc25b3_2
zstd=1.5.2=h19a0ad4_0
